+++ tests/testthat/test-combination_sampling.R
@@ -20,15 +20,72 @@ for (t in 1:T) {
     experts[t, , 2] <- qnorm(prob_grid, mean = 5, sd = 2)
 }
 
-expect_warning(boa <- online(
+expect_message(boa <- online(
     y = matrix(y),
     tau = prob_grid,
     experts = experts,
     p_smooth_pr = list(
-        knot_distance = c(0.01, 0.05, 0.5, 1),
+        knots = c(15, 20, 30, 50, 99),
         deg = c(1, 2),
         lambda = c(1:2)
     ),
     parametergrid_max_combinations = 3,
     trace = FALSE
-), "Warning: Too many parameter combinations possible. 3 combinations were randomly sampled. Results depend on sampling.")
+), "Too many parameter combinations possible. 3 combinations were randomly sampled. Results depend on sampling.")
+
+# Test if the seed works as expected
+set.seed(1)
+suppressMessages({
+    first <- online(
+        y = matrix(y),
+        tau = prob_grid,
+        experts = experts,
+        soft_threshold = c(0, 0.1, 0.2, 0.3, 0.4),
+        forget_regret = c(0, 0.1, 0.2, 0.3, 0.4),
+        parametergrid_max_combinations = 3,
+        trace = FALSE
+    )
+})
+
+set.seed(1)
+suppressMessages({
+    second <- online(
+        y = matrix(y),
+        tau = prob_grid,
+        experts = experts,
+        soft_threshold = c(0, 0.1, 0.2, 0.3, 0.4),
+        forget_regret = c(0, 0.1, 0.2, 0.3, 0.4),
+        parametergrid_max_combinations = 3,
+        trace = FALSE
+    )
+})
+
+expect_true(all(first$parametergrid == second$parametergrid))
+
+set.seed(1)
+suppressMessages({
+    first <- online(
+        y = matrix(y),
+        tau = prob_grid,
+        experts = experts,
+        soft_threshold = c(0, 0.1, 0.2, 0.3, 0.4),
+        forget_regret = c(0, 0.1, 0.2, 0.3, 0.4),
+        parametergrid_max_combinations = 3,
+        trace = FALSE
+    )
+})
+
+set.seed(2)
+suppressMessages({
+    second <- online(
+        y = matrix(y),
+        tau = prob_grid,
+        experts = experts,
+        soft_threshold = c(0, 0.1, 0.2, 0.3, 0.4),
+        forget_regret = c(0, 0.1, 0.2, 0.3, 0.4),
+        parametergrid_max_combinations = 3,
+        trace = FALSE
+    )
+})
+
+expect_false(all(first$parametergrid == second$parametergrid))